Added Bower, Datafiles and Logfiles from ENV, Added Procfile
For heroku to install npm and required node modules, package.json should be in the root folder. However, npm doesnt allow to specify a location for node modules. But Flask wants all its static files (viz. which are files from node_modules) to resides static/ folder. So moved to bower which allows us to specify a directory for static files